Cleanup: Flatten directory structure, update to new versions
This sample project had multiple subfolders, for Android Studio 2,
Android Studio 3, and Android Studio 4. Most of this was obsolete.  This
is now flattened into just one sample, which shows how to write a lint
check for a recent version of Studio/AGP.

It also updates the lint checks in various ways: using JUnit4 for the
unit test instead of JUnit3, providing a Vendor registration for the
issue registry, updating various versions of plugins and libraries to
the latest versions, etc etc.